feat: enable universal deploy
