+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@adobe/helix-resolve-git-ref",
-  "version": "1.7.17",
+  "version": "1.7.17-hedy",
   "private": true,
   "description": "An OpenWhisk Web Action that resolves a git reference (branch or tag) to the corresponding commit sha.",
   "main": "src/index.js",
@@ -12,10 +12,10 @@
     "semantic-release": "semantic-release",
     "docs": "npx jsdoc2md -c .jsdoc.json --files 'src/*.js'  > docs/API.md",
     "commit": "git-cz",
-    "build": "wsk-builder -v",
-    "deploy": "wsk-builder -v --deploy --test=/_status_check/healthcheck.json",
-    "deploy-ci": "wsk-builder -v --deploy --test=/_status_check/healthcheck.json --pkgVersion=ci$CIRCLE_BUILD_NUM -l ci",
-    "deploy-sequences": "wsk-builder --no-build -no-hints -l latest -l major -l minor"
+    "build": "hedy -v",
+    "deploy": "hedy -v --deploy=wsk --test=/_status_check/healthcheck.json",
+    "deploy-ci": "hedy -v --deploy=wsk --test=/_status_check/healthcheck.json --pkgVersion=ci$CIRCLE_BUILD_NUM -l ci",
+    "deploy-sequences": "hedy --no-build -no-hints -l latest -l major -l minor"
   },
   "wsk": {
     "namespace": "helix",
@@ -33,16 +33,17 @@
   "homepage": "https://github.com/adobe/helix-resolve-git-ref#readme",
   "dependencies": {
     "@adobe/gh-resolve-ref": "1.1.0",
+    "@adobe/helix-deploy": "1.3.0",
     "@adobe/helix-epsagon": "1.5.5",
     "@adobe/helix-status": "8.2.3",
     "@adobe/openwhisk-action-logger": "2.2.2",
-    "@adobe/openwhisk-action-utils": "4.4.0"
+    "@adobe/openwhisk-action-utils": "4.4.0",
+    "node-fetch": "2.6.1"
   },
   "devDependencies": {
     "@adobe/eslint-config-helix": "1.1.3",
     "@adobe/helix-fetch": "1.9.2",
     "@adobe/helix-log": "4.5.3",
-    "@adobe/openwhisk-action-builder": "2.15.1",
     "@pollyjs/adapter-node-http": "5.0.0",
     "@pollyjs/core": "5.0.0",
     "@pollyjs/persister-fs": "5.0.0",
